+++ packages/tauri-db-sqlite-persistence/e2e/app/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tanstack/tauri-db-sqlite-persistence-e2e-app
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies [[`8258d09`](https://github.com/TanStack/db/commit/8258d0955ab47c8510bd49ea59bcdbefd2ae054d), [`286964d`](https://github.com/TanStack/db/commit/286964d72612b59e3e427baabd9870f5a71a4281)]:
+  - @tanstack/db@0.6.16
+  - @tanstack/tauri-db-sqlite-persistence@0.2.8
+
 ## 0.0.19
 
 ### Patch Changes